4주차 권한관리 마이페이지 구현

+++ main/views.py
@@ -71,12 +71,17 @@ def secondpage(request):
     return render(request, 'main/secondpage.html', context)
 
 def new_post(request):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
     return render(request, 'main/new_post.html')
 
 def create(request):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+    
     new_post = Post()
     new_post.title = request.POST['title']
-    new_post.writer = request.POST['writer']
+    new_post.writer = request.user.username
     new_post.content = request.POST['content']
     new_post.category = request.POST['category']
 
@@ -97,13 +102,27 @@ def detail(request, post_id):
     return render(request, 'main/detail.html', {'post': post})
 
 def edit(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+    
     edit_post = get_object_or_404(Post, pk=post_id)
+
+    if edit_post.writer != request.user.username:
+        return redirect('main:detail', edit_post.id)
+    
     return render(request, 'main/edit.html', {'post': edit_post})
 
 def update(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+    
     update_post = get_object_or_404(Post, pk=post_id)
+
+    if update_post.writer != request.user.username:
+        return redirect('main:detail', update_post.id)
+    
     update_post.title = request.POST['title']
-    update_post.writer = request.POST['writer']
+    update_post.writer = request.user.username
     update_post.content = request.POST['content']
     update_post.category = request.POST['category']
     update_post.save()
@@ -111,7 +130,14 @@ def update(request, post_id):
     return redirect('main:detail', update_post.id)
 
 def delete(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+    
     delete_post = get_object_or_404(Post, pk=post_id)
+    
+    if delete_post.writer != request.user.username:
+        return redirect('main:detail', delete_post.id)
+    
     delete_post.delete()
 
     return redirect('main:postpage')
